# crypto/cipher: add optimized assembly xorBytes for ARM (NEON + non-NEON)