xorBytes is used extensively in crypto and can sometimes be the
bottleneck (e.g. CTR). 32-bit ARM is still the most common arch
for IoT, it is a slower platform so crypto optimization is important,
but unfortunately crypto/cipher does not have optimized xorBytes
for ARM.

This PR adds optimized assembly language xorBytes
implementations for both NEON and non-NEON ARM, which can
result in very large performance gains.

Example comparing current tip crypto/cipher xorBytes on
ARMv7+NEON, with results after applying this PR.

```
name                   old time/op    new time/op    delta
XORBytes/8Bytes-4         127ns ± 0%      95ns ± 0%    -25.14%
XORBytes/128Bytes-4      1.21µs ± 1%    0.16µs ± 0%    -86.70%
XORBytes/2048Bytes-4     18.4µs ± 0%     1.3µs ± 0%    -93.17%
XORBytes/32768Bytes-4     304µs ± 1%      31µs ± 1%    -89.93%

name                   old speed      new speed      delta
XORBytes/8Bytes-4      63.2MB/s ± 0%  84.4MB/s ± 0%    +33.59%
XORBytes/128Bytes-4     106MB/s ± 1%   797MB/s ± 0%   +651.92%
XORBytes/2048Bytes-4    111MB/s ± 0%  1627MB/s ± 0%  +1364.86%
XORBytes/32768Bytes-4   108MB/s ± 1%  1071MB/s ± 1%   +892.65%
```

Additionally, this PR improves xor_generic.go to use word copy
for xorBytes on all arches if the dst and src buffers are all aligned,
which in real world usage is often the case. Previously word copy
was used only on platforms that supported unaligned access,
which is unnecessarily strict and left performance on the table.

Fixes #53023